+++ trisicell/settings/__init__.py
@@ -1,5 +1,5 @@
 """Settings Module."""
 
-from trisicell.settings._settings import logfile, verbosity
+from trisicell.settings._settings import logfile, tools, verbosity
 
-__all__ = (verbosity, logfile)
+__all__ = (verbosity, logfile, tools)