+++ trisicell/tl/solver/_sasc.py
@@ -1,0 +1,3 @@
+def sasc():
+    # TODO: implement
+    return None