+++ trisicell/tl/solver/_siclonefit.py
@@ -9,6 +9,8 @@ import trisicell as tsc
 
 def siclonefit(df_input, alpha, beta, n_iters):
     # TODO: implement
+    executable = tsc.ul.executable("SiCloneFiTComplete.jar", "SiCloneFit")
+
     tsc.logg.info(
         f"running SiCloneFit with alpha={alpha}, beta={beta}, n_iters={n_iters}"
     )
@@ -23,9 +25,8 @@ def siclonefit(df_input, alpha, beta, n_iters):
         fout.write(" ".join(df_input.columns))
     I_mtr = df_input.values
 
-    siclonefit = tsc.ul.get_file("trisicell.external/bin/SiCloneFiTComplete.jar")
     cmd = (
-        f"java -jar {siclonefit} "
+        f"java -jar {executable} "
         f"-m {df_input.shape[0]} "
         f"-n {df_input.shape[1]} "
         f"-ipMat {tmpdir.name}/siclonefit.input "